+++ trunk/openvibe-plugins/signal-processing-gpl/trunc/src/box-algorithms/basic/ovpCDownsamplingBoxAlgorithm.cpp
@@ -185,7 +185,7 @@ boolean CDownsamplingBoxAlgorithm::process(void)
 
 				m_pOutputSignalDescription->setDimensionCount(2);
 				m_pOutputSignalDescription->setDimensionSize(0, m_pInputSignal->getDimensionSize(0));
-				m_pOutputSignalDescription->setDimensionSize(1, m_ui64CurrentBufferSize);
+				m_pOutputSignalDescription->setDimensionSize(1, (uint32)m_ui64CurrentBufferSize);
 				for(uint32 k=0; k< m_pInputSignal->getDimensionSize(0); k++)
 				{
 					m_pOutputSignalDescription->setDimensionLabel(0, k, m_pInputSignal->getDimensionLabel(0, k));
